backend/requirements.txt
asgiref==3.8.1
astroid==3.3.5
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
dill==0.3.9
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
identify==2.6.1
idna==3.7
isort==5.13.2
mccabe==0.7.0
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.0.0
PyJWT==2.9.0
pylint==3.3.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
tomlkit==0.13.2
urllib3==2.2.2
virtualenv==20.26.6
